Fixed bugs in Egholm benchmark topo and balance rate.
There were two bugs in the Egholm 1 benchmark:

(1) The width constant in the paper is the half-width of the grid, but was
treated as a full width here. I fixed this in the benchmarkInitGrids and
benchmarkBalanceRate subroutines.

(2) The inequalities in benchmarkBalanceRate were reversed from the equations in
the paper. I swapped these to match the paper.

I created a new benchmark case, egholm_1_ss, which runs the egholm example
without erosion until the ice is at steady state. A superficial comparison
is consistent with the paper.

Another possible issue is that the grid was defined with a very different
spacing in the x- and y-directions. This is not wrong, but there is no reason
not to run with an equal, or nearly equal, grid spacing. I fix this in the new
case.

NOTE: The sliding constant and erosion constant is not currently set for any of
the Egholm cases, and the default values (0) are used instead. This means the
ice neither slides nor erodes, which will have to be changes for the transient
cases.