Changed 'topo' back to 'grid'. Progress on 'hill' object.
I reverted the 'topo' to a 'grid' object with the idea of embedding the
interpolation subroutines in this object. It does not make sense the the
topography would be included in such an object since other variables are
interpolated as well. On top of this, it does not add much clutter to have an
independent topography variable.

For the 'hill' object, I have altered the initialization routine so that the
basic parameters are set in the io subroutine, and the vectors are allocated and
populated in the internal init routine. I have also started working on including
a steady-state benchmark case, and given some thought how the benchmarks should
be reworked in general.

The new approach is to define the forcing for all models
(benchmark and otherwise) using functions, and then to select the active
function during object initialization. The advantages are that model execution
is MUCH simplified by avoiding branches needed for special cases, and benchmark
definitions are neatly wrapped into the modules defining each object type.
Benchmarks will be run using input files, just like normal runs, and these
special input files will be included with the model source code.