Progress in implementing hill object BC procedures.
The basic implementation is in place, but is turns out that the system GCC on
the BU SCC is too old to handle procudure pointers as components of a derived
type. This means I cannot compile and test until I first rebuilt the
dependencies using a newer GCC version. This prompted me to make this premature
commit.