Fixed implementation of type-bound procedures and procedure pointers.
The procedures associated with the hill object are now properly implemented.
Type-bound procedures use the 'class' declaration for the passed dummy object,
as required by the INtel compiler. BCs are selected via a string identifier in
the input file, then are applied by associating the correct functions with the
hill object procedure pointers. THis all seems to work well, though I have not
yet debugged the steady-state benchmark case. The next step is a working
hillslope model.