Prep for full update, fix compile on SCC (using PGI)
I commented out most of the code in ice_cascade.f90, leaving only the io
sections and the new hill and grid sections enabled. I did this to simplify the
overhaul process by discarding the baggage of the previous version. The old
version, which contain lots of useful code, are moved to the subfolder
'last_version' temporarily. They will be deleted as they are fully replaced by
the new routines.

To compile on SCC, I have moved to the PGI compiler suite. I did this because
PGI supports fortran 2003 (essential for object-oriented code) but remains
compatible with most of the SCC software stack (which was compiled with GCC
4.4.7). Libraries that include .mod header files do need to be rebuilt. I did
this for netcdf-fortan/4.4.2 to start, and disabled the other dependecies for
now.