Hill boundary conditions working: zero_grad, bench_hill*
These boundary conditions are tested and working. Note: I migrated the hill
module boundary conditions to a separate module with the idea of reusing them
for the other model components. Having done this, I really don't like it - it
means the hill module is incomplete without this external module, which is not
ideal. I will be merging them back together in the next commit.