Implemented diffusion using 5-point stencil
A bit of searching on the web indicated that the 5-point stencil and 9-point
stencil both have O(dx^2, dy^2) error, except in the special case of an even
grid, where the 9-point stencil is a much better approximation. For the moment,
I use just teh simpler 5-poit stencil, but this should probably be changed in
the future since square grids are common.

The steady-state hillslope benchmark appears to be running well (kai-by-eye),
but the exact solution should be included to test this.