Cleaned up implementation of solutions, etc in hill object.
A bunch of minor changes that clean up how the object is used. These include
adding a dzdt component to the hill object, converting the solution procedure to
a function, properly clearing all components when the object is set to off, etc.
It appears that the benchmark is still working as expected.