Bugfixes needed to compile and run using GCC 4.4.7 on the BU SCC
1) The ieee_arithmetic module used in ice.f90 to a default timeStep value of
infinity is not supported in older compilers. As an alternative, I use huge() to
set the largest available number. This should have the same effect of ignoring
timesteps computed where there is no ice.

2) Calls to mpi_waitall throw errors when the request values are not set. The
newly implemented 'if proc .ne. mpi_proc_null' statements cause this problem. A
simple solution is to add a 'then' statement to set a null request value.