Created new climate object module, simplified others.
The new climate module is comlete up through initialization - the run procedure
is not yet defined and the variables are not yet written to the output.

In the process of writing this module, I found that it is simpler to initialize
the coordinate variable in each model component using the grid_type object as a
init() subroutine argument. This has two benefits, 1) the grid associated with
each object is clearer in ice_cascade.f90, regardless of how I name the
variables, and 2) there is no need to copy over gird definition variables to all
the objects in io.f90, this is handled in thier init() routines instead.