Writes climate header, bugfix for disabled climate model.
The parameters that define the climate model are now written as global
attributes, if the model is enabled.

I fixed a bug where the program crashed if the clamate model was disabled. The
problem was that the run() subroutine ws called even for a disabled model, which
attempted to access a NULL pointer and threw a segmentation fault.