Added surface ice flux calculation to the climate object.
Added new variables and procedures to the climate object to compute surface ice
flux (a.k.a. balance rate). I had originally planned to include this in the ice
model object instead, but realized that ice flux models may or may not depend on
temperature, precipitation, or other climate variables not currently included,
it makes more sense to have the ice flux closely associated with other climate
variables.

Currently, surface ice flux variables are initialized, but not computed or written.