Minor update before starting on ice model component
Bugfix: The main program attempted to use the 'on' component in the climate
object, which I had removed. I re-added this to the object for convenience,
which fixed the problem. I also combed the code for places where the re-added
member could allow for simplification.

Update: I renamed the Makefile for Yale Omega, and now set the default Makefile
using a symbolic link. This is just more convenient.