Added exact solution for Bueler isothermal test A.
This benchmark is symmetric, so I plan to compute it for the northeast quadrant
only. The paper sets two paramters for test A: L = 750km, the half-width of the
ice cap, and M0 = 0.3 m/a, the constant accumulation rate. In my model, L is set
by defining the number of grid points and grid spacing such that the domain is
750km wide, and M0 is set as a hardcoded parameter in ice.f90