Restructuring IO: wrote python tools to generate input files.
After finding the original input method (reading a text file for parameters, and
reading initital grids from separate netCDF files) to be too inflexible to deal
with even the benchmark cases, I decided to make the input files share the
format of the output format. This means input files are netCDF files with all
model parameters as global attributes, and gridded data as variables. A key
difference between output and input files is that there is no time dimension in
the inputs. This means that the user would have to extract a single timestep to
use an output file as a starting point for a new model run.

To make this easier, I wrote a python module with to create a netcdf with all of
the variables expected by the IO routine. The way to build a new model run is
to via a python script that makes use of this toolbox. THis will make it simple
to add new variables as needed, without braking existing generation scripts.

The IO routine is NOT yet updated. This is the next step.