Started work on the bueler_isothermal_a method.
I decided that 'sets' of methods like this can be stored in separate modules.
This allows for two useful things: (1) the ability to group related methods for
different model components (e.g. climate and solution for the ice benchmarks),
and (2) the ability to include an 'init' function to precompute expensive
constant variables (for example the radial distance).