New ice flow method, and removed padding from state_type vars
The new ice flow method is in mtd_hindmarsh2_explicit, so far I have only
written the skeleton of the module, the update subroutine is not yet
implemented.

The state vars now contain only interior points (i.e. no ghost points used for
boundary condition treatments). The reason for this change is that I realize
that there are some methods that may required more than one ghost point at the
margins (e.g. MUSCL), and thus that the program was less flexible than it may
need to be by padding the shared variables in a state_type object. The
alternative is to store only the actudal data in state_type, and make padded
copies as needed in the subroutines. The cost for doing this should be small (a
few copy operations upon entering the method subroutines, every big model step)
and it has the benefit of simplifying the IO code and clarifying the treatment
of boundaries by explicitly padded in each subroutine as needed.