Check-in prior to major change
I realized an unfortunate flaw in my program design: the modular design has
circular dependencies, which is not allowed in fortran. These arose because the
ice model methods are needed in the ice module, but these methods need ice_type
from the same module. I am still working out a solution for this problem.