Simplified python tools.
Instead of creating and passing around a dictionary constaining the parameters
and state variables, I lean more heavily on the netCDF4 library and simply
create the file and pass around the open dataset object for modification. The
result is to remove about half the code from the python tools, and completely
eliminate the need for the user to remeber the dictionary keys.