Added run script for bueler test A, also found a problem.
The problem is that the way written time steps are specified is very fragile -
they must be integer multiples of the model timestep. This is hard to deal with
for two reasons. First it is necessary to be very careful with floating point
number comparisons. Second, it will not be possible to extend this to deal with
adaptive model timesteps, which I would like to add in the distant future.

An alternative is to specify the output timesteps as a vector in the input file.
This will be quite flexible and reasonable easy to implement.