Revised output timestep specification.
Output timesteps are now specified as a vctor of specific times when the output
sould be written. The only restrictions are that the vector must be monotonic
(increasing for forward models, and decreasing for reverse models) and within
the limits of the start and end times. The reverse case is not yet fully
tested.