Tested bueler_isothermal_b and fixed related bugs
Many small bugs related to the specific test case were fixed.

One major bug in the main model was fixed. This was that the ice model update
subroutine used the (constant) param%time_step for the model time step, which
the main routine has a variable timestep to accomodate writing output at
arbitrary times. The solution was to include the variable time step in the state
object.