Identified source of excessive error in bueler_isothermal_c tests
The problem was that the model time step was too large, so that the exact
surface ice flux and model surface ice flux were substantially different.
Decreasing the model time step from 100a to 1a brought the results into line
with published errors.

This result strongly suggests that the climate computation should be moved into
the ice model loop (provided that the ice model is active). I have decided not
to do this yet, but will likely do so in the near future.