New utility function to read netcdf vars, plus a bugfix
The new function get_var_2 in io.f90 reads 2D netcdf variables, handles errors
returned by the netcdf library, and omits variables that are null (all NaN). It
is tested in a limited way and appears to be working correctly.

Bugfix: The climate_bueler_isothermal_a intialization function was missing an
allocation section, which caused the program to fail. This is fixed.