Finished bueler_isothermal_e climate module
Finished the climate_bueler_isothermal_e module and carefully checked the
results against the paper. There were several problems and typos in the previous
version that needed to be fixed to get everything working correctly. Most
notably, the equations for compensatory accumulation are undefined at radius =
0. Computing the value at r=0 lead to unused, but unnecessary Inf values in the
grid variables. This also occured in the ghost points, for which I had left the
x and y coordinate values as 0. The solution was to compute compensatory
accumulation only in the sliding region (where it is needed). Though not
strictly necessary, I also updated the IO module to define the x- and y-
coordinates of the ghost points, to avoid unnecessary floating point exceptions
in the future.

I left my debugging code in for this commit, and will remove it in the next.