Created ice_flow_hindmarsh2_sliding_explicit module
Currently contains only the header with a description of the module and the
method.

Note: For this module, I plan to read the state variables ice_defm_a and
ice_slid_a to get the deformation and sliding parameters. Thus the method will
have no input parameters, but can fail if these state variables are not set
during intialization (or potentially elsewhere).

I plan to make this change to the original hindmarsh2_explicit module as well.
It makes sense to use these variables and permit users to have a deformation
constant that is not constant in space.