Flow module in progress.
The new hindmarsh2_sliding_explicit module is in progress. The major differences
are (1) the deformation and sliding coefficients come from grids, and so it is
necessary to use midpoint values on the staggered grids, and (2) I make no
pretense at efficiency, but rather attempt to code as clearly and correctly as
possible. My intention here is to get to a working copy as quickly as possible,
then to improve speed by iterative profiling and correction.

So far, I have updated only the x-direction ice flux. It compiles, but until the
work is done, I have no way to check for bugs.