Updated no_ice BC so that it includes the coefficient grids.
Boundary points for the ice_a_defm and ice_a_slid grids must be populated for
the BCs to work as planned. This new version of the no_ice boundary conditin
sets the ghost points for the coefficient grids equal to the adjacent interior
points. I cannot see a good reason not to do it this way, but am not totally
confident that it is correct either.