Updated no_ice BC so that it includes the coefficient grids
Interestingly, no substative change is needed. The ice flux on the boundarys
goes to zero because the surface gradient goes to zero, regardless of the value
of the coefficients. For this reason, the BC works find as is. This update
simple changes the comments to more clearly reflect this fact.