Accuracy problem diappeared. Fixing grid refinement bug.
The grid refinement experiment refuses to run for bueler_isothermal_e and stderr
is not currently printed by the subprocess.call(). I attempted to run the
program in python 3.4 (which would make printing stderr easy) but this created a
new set of problems (netcdf files with "string" type attributes that are not
supported in the Fortran 90 netCDF API).

Next steps are to get stderr to print in python 2.7, then diagnose and fix the
problem with the bueler_isothermal_e input files.