Fixed grid refinement bug for bueler_isothermal_e
Bug description: The grid refinement experiment python script failed for test
bueler_isothermal_e without error. A new version of the bueler_isothermal_a test
using the new hindmarsh2_sliding_explicit flow method ran fine, so the problem
appeared to be with the python scripts. Oddly, the model ran successfully using
the input files generated by the grid refinement script descpite having just
failed. Using print statements in the ice-cascade source, I found that all the
variables were being read in as NaN, although ncdump was able to find them
correctly.

Bug fix: The bug was caused by a omitted close statement finalizing the netcdf
file in the input_bueler_isothermal_e.py script. The fix was simply to add this
statement. My best guess at the cause for the symptoms is that the data was not
written to the input netcdf file until it was closed. I had not closed the file
explicitly, but it is likely closed implicitly when the parent python script
terminates and the file object is destroyed. Thus, when running from within the
grid refinement script, I was using an incomplete (open) file for input data,
and when I ran afterwords, the file had been closed and everything worked well.

Changes: There are changes all over the place, almost all of them made as part
of debugging.