Switching to test_* files with common arguments
Migrated the code for parsing input arguments (which is common to all of the
test cases) to the ice_cascade_tools module. The old input_* utilities will be
replaced with an equivalent set of test_* utilites that make use of this
function for parsing thier input arguments, and allow for switching the active
ice flow method.