Updated bueler_isothermal_c, fixed associated bugs
The updated test_bueler_isothermal_c.py utility lead to floating-point-errors in
the ice-cascade run. These turned out to be caused by the fact that the model
started at time = 0, which meant that ice thickness was also zero. This caused
several divide-by-zero errors, in both the climate_bueler_isothermal_c routine,
and also in the hindmarsh2_explicit routine.

I fixed the problem by starting the model at an arbitrarily small time greater
than zero. One side-effect is that I was able to remove unnecessary (and
ineffective) protections from the climate routine.

NOTE: This bug points out an important flaw in my program - it will fail with a
floating point error if the maximum ice diffusivity is zero, which will be the
case if ice thickness is uniformly zero. This bug needs to be fixed, in some
inexpensive way.