Restructured directories and added the setup_environ.sh tool
Directory structure changes are intended to better separate the utility python
scripts based on thier purpose. To do this I created new directories for test
scripts and for experiments. The experiments directory can hold user's
experiment setup scripts, and should grow into a useful set of examples as more
people use the package.

The new setup_environ.sh utility is intended to help user's automatically setup
thier shell environment for ice-cascade and the associated utilities. I expect
the requisite commands to be different for everyone, so the user can add a new
system_name and associated commands as needed.