Added convenience function for reading netcdf vars to ice_cascade_tools
The goal here was to write a function that takes care of reading and checking a
netcdf variable. Having written it, I now think this is unnecessary - most of
the function consists of checking the function arguments, and the basic
functionality boils down to just a few lines. I will delete this in the next
commit, but leave it here in case I decide to return to this path in the future.