Bugfix: eliminated unnecessary ice velo computations
The program was computing ice velocity for every timestep, it is only needed for
timesteps when output is written.

The solution was to make the update_ice subroutine aware of when output will be
written by passing in a logical flag.

Note that this will also allow the ice exact solution computation to be isolated
within the ice module (in a future commit).