Moved solve_ice call from main to ice module
This simplification was made possible by the addition of a write_now flag to the
update_ice argument list. Other minor changes include adding a new parameter
sanity test to bueler_isothermal_a and _b modules, which throws an error if a
solution is defined without setting write_ice_h_soln to 1. This combination of
varables is absurd and should be prevented. These changes need to be propagated
through the stack of tests.

Compiles, and appears to run fine with bueler_isothermal_a, but fails with a FPE
for bueler_isothermal_b.