Renamed the sliding ice flow method to sliding1
The motivation for the change was realizing that I had coded the sliding method
for bueler_isothermal_e, which uses a linear relation between sliding velocity
and basal shear stress, and the fact that most glacial models use a power law
relation where sliding velocity scales with the cube of basal sliding stress.
Thus, I will need to make a new sliding method and need to differentiate
between the two.